added relationtests
