+++ test/custom_methods.test.ts
@@ -1,4 +1,3 @@
-// tslint:disable-next-line:no-implicit-dependencies
 import test from 'ava';
 
 import { nohm } from '../ts';
